laba 3 criteria api and metadata api

+++ src/main/java/dao/ClientDAO.java
@@ -5,8 +5,8 @@ import model.Client;
 
 public interface ClientDAO {
 
-    Integer createClient(Client client) throws DAOException;
+    Long createClient(Client client) throws DAOException;
 
-    Client getClientById(Integer id) throws DAOException;
+    Client getClientById(Long id) throws DAOException;
 
 }
